~~ WIP ~~ Switch all current 2.2 versions to 2.3
`toChar()` is marked as `errorSince = "2.3"`.

+++ compiler/testData/codegen/box/primitiveTypes/kt14868.kt
@@ -2,5 +2,5 @@
 fun box(): String {
     val x: Number = 75
 
-    return "O" + x.toChar()
+    return "O" + x.toInt().toChar()
 }
