+++ compiler/testData/codegen/box/primitiveTypes/numberToChar/simple.kt
@@ -13,6 +13,6 @@ class MyNumber(val value: Int) : Number() {
 }
 
 fun box(): String {
-    val x = MyNumber('*'.code).toChar()
+    val x = MyNumber('*'.code).toInt().toChar()
     return if (x == '*') "OK" else "Fail: $x"
 }
